Initial set of Charts to deploy the Alvearie Clinical Ingestion Pattern
This commit includes an initial set of charts to deploy the Alvearie
Clinical Ingestion pattern.

There is a main chart that deploys the whole framework (including its
components) and an extra couple of charts used to deploy IBM FHIR and a
NiFi Registry backed by a Github repository.

The Alvearie Clinical Ingestion pattern chart deploys the following
related services:

- IBM FHIR: A FHIR server to store medical data
- Kafka: For clinical ingestion messaging
- NiFi: To deploy the ingestion flows
- NiFi Registry: To deliver the ingestion flows to be used in NiFi
- Zookeeper: Dependency needed by Kafka & NiFi

The corresponding documentation with instructions is included.

The added documentation pages are pretty standalone. There is a sentence
reference in the main clinical-ingestion/README.md that leaves the
existing instructions as-is for those who may already have the service
dependencies, and introduces and references the Helm Chart for those who
don't.