Enhance the helm chart to include the new monitoring stack
Importing the chart for kubernetes/prometheus/grafana. Added configuration to allow kafka
and nifi ports with associated scrape configs for both.  Custom grafana dashboards are included
for nifi and kafka as well as a selection of three kubernetes monitors.  Finally, a script is
used to automate the configuration of the reporting task for the nifi port.  The script is executed
after first checking that the nifi cluster is active and ready to accept requests.