Integrate new deid chart with the alvearie-ingestion chart
This commit includes the deid chart created in a previous commit with
the alvearie-ingestion chart. By default the deid flow services
(secondary FHIR and deid service) are disabled, but a second
values file (`de-id-pattern-values.yaml`) with them enabled is included
for users wanting to use the de-id pattern.

This commit also includes a default _helpers.tpl that will likely be
necessary for a future task on writing the NOTES.txt file for the
alvearie  chart.