Add connectivity tests to the various charts
This commit adds basic connectivity tests to the various charts that we
currently have:

- alvearie-ingestion
- deid
- fhir
- nifi-registry

The tests are added individually to each subchart, and to the main
chart. When a test is added to a subchart, such as fhir, that test can
be executed on a standalone chart deployment of that specific subchart.

Additionally, tests for subcharts (ours and the external ones we use)
are automatically executed when the main chart is tested. Therefore for
the main chart, I only added a connectiviry test for the nifi chart,
since that chart does not include any standalone tests.

The tests are run as follows:

```
helm install ingestion .
helm test ingestion
```