Add Chart linting workflow
This commit adds the workflow definition and the necessary changes for
the chart linting process to succeed.

Those changes are:

- Add a new Github Action called `chart-testing-action` that runs `ct
lint` over the changed chart directories
- Move subcharts under the `charts/` directory: The reason this is
needed is that the `helm lint` command does not recognize `tgz` chart
packages, it can only download the dependent charts from a repository or
introspect an actual chart directory.
- Fix multiple YAML and Helm lint warnings