Implement CI Chart install
This commit enhances the existing charts CI pipeline to include
installing the chart in addition to just linting the charts.

The installation is done in a `kind` kubernetes cluster that is setup as
part of the Github Action.

A new `*-values.yaml` file is added that has the configuration necessary
to install the chart for testing purposes. That includes disabling
certain components that are not supported by the `kind` cluster such as
the prometheus stack and using `LoadBalancer` services.

This test values.yaml also includes installing the de-id service and the
FHIR Proxy for better coverage.

Also, provide a specific FHIR Proxy container image version.