Add Cohort Service sub-chart to Clinical Ingestion Chart
This commit adds the Cohort Service's chart as a sub-chart of the
Clinical Ingestion main chart. This allows users to try out the cohort
service if needed. By default the service is disabled.

This commit also enabled the `LoadBalancer` for the FHIR service by
default per a discussion in slack about why we had turned it to
`ClusterIP`, which may have been accidentally.

Since we added a new service I bumped the chart version a whole point :)

After this commit the ingestion framework may include the cohort service
as shown below:


```
Welcome to the

       d8888 888                                    d8b
      d88888 888                                    Y8P
     d88P888 888
    d88P 888 888 888  888  .d88b.   8888b.  888d888 888  .d88b.
   d88P  888 888 888  888 d8P  Y8b     "88b 888P"   888 d8P  Y8b
  d88P   888 888 Y88  88P 88888888 .d888888 888     888 88888888
 d8888888888 888  Y8bd8P  Y8b.     888  888 888     888 Y8b.
d88P     888 888   Y88P    "Y8888  "Y888888 888     888  "Y8888

                               Clinical Data Ingestion Framework

Pattern: Ingestion

The following services have been deployed:
- Primary FHIR Server
- NiFi Server
- NiFi Registry
- Zookeeper
- Cohort Service
```