Refactor Common Component with base FHIR Service
The de-id and terminology service use cases both require communication
to the FHIR Server and they had several similarities in terms of both
their implementation and testing.

This commit implements a refactoring extracting the common code into a
new couple of classes FHIRService and FHIRServieTest, that live in a new
project called common.

Refactoring the tests of the custom processors to extract commonalities
is also included in this commit.

The corresponding necessary build changes to build and get the right
visibility on the child projects are included too.