Update Cohort Engine dependencies in the Cohort Service App
A recent update by the Cohort Engine team changed some public packages
that we are using in the Cohort Engine app. This commit updates those
pacakges so the app builds against their latest version of the code.

Additionally, their Github repository is added to the POM so that the CI
build will pass.