Fix classpath issues
There are some HAPI FHIR JARs that dependend on the same upstream OSS
JARs that are required for cohorting. HAPI uses version 1.4.9 and
cohorting uses version 1.5.1. We were getting part of one version and
part of the other which caused CQL translation errors at runtime. This
commit pins the upstream info.cqframework JARs to the newer 1.5.1
version that the cohort-engine requires.