Bug fixes for cohort-engine 1.0.1 pre-release updates (#202)
* Fix classpath issues

There are some HAPI FHIR JARs that depend on the same upstream OSS
JARs that are required for cohorting. HAPI uses version 1.4.9 and
cohorting uses version 1.5.1. We were getting part of one version and
part of the other which caused CQL translation errors at runtime. This
commit pins the upstream info.cqframework JARs to the newer 1.5.1
version that the cohort-engine requires.

* Add a call to refreshCQLDirectory on updateLibrary

The DirectoryBasedLibrarySourceProvider caches directory contents on
initialization. Whenever the directory changes, a new library source
provider needs to be created. Added a call to the restCQLDirectory()
method in the updateLibrary call so that things get reiinitialized.