Adjust which CqlEvaluator evaluate method is utilized.
There are two types of callbacks available to clients using the
CQLEvaluator. One is a legacy callback that only exposes expression
complete events and the other is a newer API that exposes context
begin/end events in addition to expresion complete events. The
CohortService code was using an API that accepts the legacy callback,
but passing a modern callback. CqlEvaluator will wrap a legacy callback
in a modern callback proxy and noop the context begin/end events in
those cases. CohortService needs the context begin/end events, so I
updated the code to use the the correct API call for an
EvaluationResultCallback.